Practica finalizada

+++ tests.sh
@@ -2,6 +2,7 @@ TESTSPORO=25
 TESTSVECTOR=19
 TESTSTLISTA=15
 TESTSABBPORO=23
+TESTSAVLPORO=13
 nombreBase="tad"
 extension=".cpp"
 zero="0"
@@ -11,10 +12,11 @@ echo "1. Test TPoro"
 echo "2. Test TVectorPoro"
 echo "3. Test TListaPoro"
 echo "4. Test TABBPoro"
-echo "5. Todos"
+echo "5. Test TAVLPoro"
+echo "6. Todos"
 echo 'Opcion: '
 read Opcion
-if (( $Opcion == 1 || $Opcion == 5 )); then
+if (( $Opcion == 1 || $Opcion == 6 )); then
 	
 	echo "//////////////////////////////////////////////////////////////////"
 	echo "///////////////////////Pruebas TPoro//////////////////////////////"
@@ -39,7 +41,7 @@ if (( $Opcion == 1 || $Opcion == 5 )); then
 	done
 	rm src/*.cpp
 fi
-if (( $Opcion == 2 || $Opcion == 5 )); then
+if (( $Opcion == 2 || $Opcion == 6 )); then
 
 	echo "//////////////////////////////////////////////////////////////////"
 	echo "//////////////////////Pruebas TVectorPoro/////////////////////////"
@@ -64,7 +66,7 @@ if (( $Opcion == 2 || $Opcion == 5 )); then
 	done
 	rm src/*.cpp
 fi
-if (( $Opcion == 3 || $Opcion == 5 )); then
+if (( $Opcion == 3 || $Opcion == 6 )); then
 	echo "//////////////////////////////////////////////////////////////////"
 	echo "///////////////////////Pruebas TListaPoro/////////////////////////"
 	cp src/tlistaporo/*.cpp src/
@@ -88,7 +90,7 @@ if (( $Opcion == 3 || $Opcion == 5 )); then
 	done
 	rm src/*.cpp
 fi
-if (( $Opcion == 4 || $Opcion == 5 )); then
+if (( $Opcion == 4 || $Opcion == 6 )); then
 	echo "//////////////////////////////////////////////////////////////////"
 	echo "///////////////////////Pruebas TABBPoro///////////////////////////"
 	cp src/tabbporo/*.cpp src/
@@ -104,6 +106,30 @@ if (( $Opcion == 4 || $Opcion == 5 )); then
 		make > basura.txt
 		./tad > $i.txt
 		diff -b $i.txt salidas/tabbporo/$nombre.sal
+		mv src/tad.cpp src/$nombre
+		rm ./tad
+		rm ./basura.txt
+		rm $i.txt
+		
+	done
+	rm src/*.cpp
+fi
+if (( $Opcion == 5 || $Opcion == 6 )); then
+	echo "//////////////////////////////////////////////////////////////////"
+	echo "///////////////////////Pruebas TAVLPoro///////////////////////////"
+	cp src/tavlporo/*.cpp src/
+	for (( i = 1; i <= $TESTSAVLPORO; i++ )); do
+		
+		if (($i >= 10)); then
+			nombre=$nombreBase$i$extension 
+		else
+			nombre=$nombreBase$zero$i$extension 
+		fi
+		echo $nombre
+		mv src/$nombre src/tad.cpp
+		make > basura.txt
+		./tad > $i.txt
+		diff -b $i.txt salidas/tavlporo/$nombre.sal
 		mv src/tad.cpp src/$nombre
 		rm ./tad
 		rm ./basura.txt
